requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Model-specific dependencies (install as needed)
# OpenAI
langchain-openai>=0.0.2
openai>=1.3.0

# Google Gemini
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0

# Mistral AI
langchain-mistralai>=0.0.2
mistralai>=0.0.9

# Ollama (local models)
langchain-community>=0.0.16